+++ mmdet/core/export/__init__.py
@@ -1,3 +1,4 @@
+from .model_wrappers import ONNXRuntimeDetector
 from .onnx_helper import (add_dummy_nms_for_onnx, dynamic_clip_for_onnx,
                           get_k_for_topk)
 from .pytorch2onnx import (build_model_from_cfg,
@@ -5,10 +6,7 @@ from .pytorch2onnx import (build_model_from_cfg,
                            preprocess_example_input)
 
 __all__ = [
-    'build_model_from_cfg',
-    'generate_inputs_and_wrap_model',
-    'preprocess_example_input',
-    'get_k_for_topk',
-    'add_dummy_nms_for_onnx',
-    'dynamic_clip_for_onnx',
+    'build_model_from_cfg', 'generate_inputs_and_wrap_model',
+    'preprocess_example_input', 'get_k_for_topk', 'add_dummy_nms_for_onnx',
+    'dynamic_clip_for_onnx', 'ONNXRuntimeDetector'
 ]
